Correct ServerRequest::getParam method PHPDoc
The PHPDoc type of ServerRequest class getParam method's $default argument
is changed from "string" to "mixed", as dependent on request parameter its
default value can be expected to have another type(e.g. boolean). And if
there is a PHPStan or other analyser running in project an error is reported on
cases, where $default argument type is not string.  So as there is no obstacle for
doing this change, it is correct to change PHPDoc type of the argument.