Fix the composer provide rule
This package does not provide the code of the psr/http-factory package (which defines interfaces). It provides psr/http-factory-implementation which is the virtual package to represent implementations of the interface.
Providing the wrong package while also requiring it creates issues with Composer 2, because the solver will consider that install psr/http-factory is not necessary as it is already provided.